CAUSEWAY-3409: simplify hierarchy of DomainEventFacetAbstract
