+++ core/metamodel/src/main/java/org/apache/causeway/core/metamodel/facets/properties/property/modify/PropertyClearFacetForDomainEventFromPropertyAnnotation.java
@@ -32,7 +32,7 @@ implements PropertyClearFacet {
             final Class<? extends PropertyDomainEvent<?, ?>> eventType,
                     final PropertyOrCollectionAccessorFacet getterFacet,
                     final PropertyClearFacet clearFacet,
-                    final PropertyDomainEventFacetAbstract propertyInteractionFacet,
+                    final PropertyDomainEventFacet propertyInteractionFacet,
                     final FacetHolder holder) {
 
         super(PropertyClearFacet.class, eventType, getterFacet, null, clearFacet, propertyInteractionFacet, holder);
