+++ core/metamodel/src/main/java/org/apache/causeway/core/metamodel/facets/properties/property/modify/PropertyDomainEventFacet.java
@@ -18,14 +18,138 @@
  */
 package org.apache.causeway.core.metamodel.facets.properties.property.modify;
 
+import java.util.Optional;
+import java.util.function.BiConsumer;
+
+import org.apache.causeway.applib.events.domain.AbstractDomainEvent;
+import org.apache.causeway.applib.events.domain.PropertyDomainEvent;
+import org.apache.causeway.applib.services.i18n.TranslatableString;
+import org.apache.causeway.applib.services.i18n.TranslationContext;
+import org.apache.causeway.applib.services.i18n.TranslationService;
+import org.apache.causeway.commons.internal.base._Casts;
+import org.apache.causeway.core.metamodel.consent.Consent.VetoReason;
+import org.apache.causeway.core.metamodel.facetapi.FacetHolder;
+import org.apache.causeway.core.metamodel.facets.DomainEventFacetAbstract;
+import org.apache.causeway.core.metamodel.facets.DomainEventHelper;
+import org.apache.causeway.core.metamodel.facets.propcoll.accessor.PropertyOrCollectionAccessorFacet;
 import org.apache.causeway.core.metamodel.interactions.DisablingInteractionAdvisor;
 import org.apache.causeway.core.metamodel.interactions.HidingInteractionAdvisor;
+import org.apache.causeway.core.metamodel.interactions.ProposedHolder;
+import org.apache.causeway.core.metamodel.interactions.UsabilityContext;
 import org.apache.causeway.core.metamodel.interactions.ValidatingInteractionAdvisor;
+import org.apache.causeway.core.metamodel.interactions.ValidityContext;
+import org.apache.causeway.core.metamodel.interactions.VisibilityContext;
+import org.apache.causeway.core.metamodel.object.ManagedObject;
+import org.apache.causeway.core.metamodel.specloader.specimpl.OneToOneAssociationMixedIn;
 
-/**
- * Corresponds to <tt>@Property(domainEvent=...)</tt> annotation in the Causeway programming model.
- */
-public interface PropertyDomainEventFacet
-extends HidingInteractionAdvisor, DisablingInteractionAdvisor, ValidatingInteractionAdvisor {
+public class PropertyDomainEventFacet
+extends DomainEventFacetAbstract<PropertyDomainEvent<?, ?>>
+implements
+    HidingInteractionAdvisor,
+    DisablingInteractionAdvisor,
+    ValidatingInteractionAdvisor {
+
+    private final DomainEventHelper domainEventHelper;
+
+    private final PropertyOrCollectionAccessorFacet getterFacetIfAny;
+    private final TranslationService translationService;
+    private final TranslationContext translationContext;
+
+    /**
+     * @param getterFacetIfAny - will be null if this is for a mixin {@link OneToOneAssociationMixedIn}.
+     */
+    public PropertyDomainEventFacet(
+            final Class<? extends PropertyDomainEvent<?, ?>> eventType,
+            final EventTypeOrigin eventTypeOrigin,
+            final PropertyOrCollectionAccessorFacet getterFacetIfAny,
+            final FacetHolder holder ) {
+
+        super(PropertyDomainEventFacet.class, eventType, eventTypeOrigin, holder);
+        this.getterFacetIfAny = getterFacetIfAny;
+
+        this.translationService = getTranslationService();
+        this.translationContext = holder.getTranslationContext();
+
+        domainEventHelper = DomainEventHelper.ofServiceRegistry(getServiceRegistry());
+    }
+
+    @Override
+    public String hides(final VisibilityContext ic) {
+
+        final PropertyDomainEvent<?, ?> event =
+                domainEventHelper.postEventForProperty(
+                        AbstractDomainEvent.Phase.HIDE,
+                        _Casts.uncheckedCast(getEventType()), null,
+                        getFacetHolder(), ic.getHead(),
+                        null, null);
+        if (event != null && event.isHidden()) {
+            return "Hidden by subscriber";
+        }
+        return null;
+    }
+
+    @Override
+    public Optional<VetoReason> disables(final UsabilityContext ic) {
+
+        final PropertyDomainEvent<?, ?> event =
+                domainEventHelper.postEventForProperty(
+                        AbstractDomainEvent.Phase.DISABLE,
+                        _Casts.uncheckedCast(getEventType()), null,
+                        getFacetHolder(), ic.getHead(),
+                        null, null);
+        if (event != null
+                && event.isDisabled()) {
+
+            final TranslatableString reasonTranslatable = event.getDisabledReasonTranslatable();
+            final String reasonString = reasonTranslatable != null
+                    ? reasonTranslatable.translate(translationService, translationContext)
+                    : event.getDisabledReason();
+
+            return Optional.ofNullable(reasonString)
+                .map(VetoReason::explicit);
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public String invalidates(final ValidityContext ic) {
+
+        if(getterFacetIfAny == null) {
+            return null;
+        }
+
+        // if this is a mixin, then this ain't true.
+        if(!(ic instanceof ProposedHolder)) {
+            return null;
+        }
+        final ProposedHolder ph = (ProposedHolder) ic;
+
+        final Object oldValue = getterFacetIfAny.getProperty(ic.getTarget(), ic.getInitiatedBy());
+        final ManagedObject proposedAdapter = ph.getProposed();
+        final Object proposedValue = proposedAdapter != null ? proposedAdapter.getPojo() : null;
+
+        final PropertyDomainEvent<?, ?> event =
+                domainEventHelper.postEventForProperty(
+                        AbstractDomainEvent.Phase.VALIDATE,
+                        _Casts.uncheckedCast(getEventType()), null,
+                        getFacetHolder(), ic.getHead(),
+                        oldValue, proposedValue);
+        if (event != null && event.isInvalid()) {
+            final TranslatableString reasonTranslatable = event.getInvalidityReasonTranslatable();
+            if(reasonTranslatable != null) {
+                return reasonTranslatable.translate(translationService, translationContext);
+            }
+            return event.getInvalidityReason();
+        }
+
+        return null;
+    }
+
+
+    @Override
+    public void visitAttributes(final BiConsumer<String, Object> visitor) {
+        super.visitAttributes(visitor);
+        visitor.accept("getterFacet", getterFacetIfAny);
+    }
 
 }
