+++ core/metamodel/src/main/java/org/apache/causeway/core/metamodel/facets/properties/property/modify/PropertySetterOrClearFacetForDomainEventAbstract.java
@@ -65,7 +65,7 @@ implements
                     final PropertyOrCollectionAccessorFacet getterFacet,
                     final PropertySetterFacet setterFacet,
                     final PropertyClearFacet clearFacet,
-                    final PropertyDomainEventFacetAbstract propertyDomainEventFacet,
+                    final PropertyDomainEventFacet propertyDomainEventFacet,
                     final FacetHolder holder) {
 
         super(facetType, eventType, holder);
